requirements.txt
# =============================================================================
# Document OCR Pipeline - Python Dependencies
# =============================================================================

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Document processing - Native extraction
pymupdf>=1.23.0
openpyxl>=3.1.0
pandas>=2.0.0
python-docx>=1.0.0

# OCR - LightOnOCR
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
pypdfium2>=4.0.0

# AWS (optional - for Textract)
# boto3>=1.28.0

# Utilities
requests>=2.31.0
